==> AQUATOX/Code/R_Analyses/01_Model_Validation_Control/2021_Validation/07_Macroinvertebrates_conversion_to_biomass_2021.Rmd <==
---
title: "Convert Macroinvertebrate Abundances to Biomass"
author: "A. Lindborg"
date: "11/30/2022"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Purpose
Convert macroinvertebrate control data from abundance to biomass, using recent length-weight information provided by GmbH.

## Steps
1) Determine biomass per individual
    - Use the "regression midpoints" values provided by Jurgen/GmBH, on the "Macroinvertebrates" tab in the "Additional Parameters Mesocosm 211210_vs.xlsx" worksheet.
    - We are using the regression midpoints, and not the populations data, because that is what the other modeling groups are using for biomass conversions (i.e. model harmonization)
    - Assign each group it's length/weight (LW) value. 
    - If a specific species doesn't have LW values, assign the LW values of the closest proximity (e.g. Odonates, species Aeshna cyanea is a type of Anisoptera, so assign Anisoptera values)
    
2) Convert control data measurements from abundance (individuals/sample) to biomass (g/m2 or g/L)
    - Use the conversion factors and equations from Waterborne 
  
3) Sum the total biomass for the modeled group
    - For multi-species groups (e.g. Gastropods), determine biomass for each of the species (above step), but sum together for 1 gastropod biomass value
    
    



# Documentation
Relevant files in the documentation folder calibration analyses:

"Email_Decision_on_Calibration_Biomass_Conversion_2021-12-10.pdf"
"Email_Updated_Length_Data_2021-12-10.pdf"

# History
11/30/2022 - A. Lindborg creates script
4/28/2023 - A. Lindborg updates to include full 2021 dataset

# Setup
```{r setup}
library(tidyverse)
library(openxlsx)
library(janitor)
library(RColorBrewer)


options(scipen = 9999)
```

# Data
This analysis uses 3 data sets:

1. Macroinvertebrate LW measurements collected by GmbH 
  - "Populations" data that has average lengths of some species/groups by date (avg_length_mm). These data come from mesocosm studies (although NOT the same mesocosm studies we are modeling) -- Will need to use this to calculate an indiviudal weight for Lymnaeidae <0.5cm
  - Midpoint values for all groups, based on separately collected data (these data come from a real pond environment)
  - Parameters for study specific allometric equations
  
2. Conversion factors from Waterborne:
  - These were determined from previous meetings and are for converting individuals/sample to individuals/area for the Mesocosm control data
  
3. Mesocosm control data from Waterborne
  -abund_sum_lifestage = sample abundances, with lifestages rolled up into 1 value (e.g. can include larvae, pupae, adults)


Important Data Notes:

-Weight data that are provided in the "Macroinvert_All" sheet are in mg-- this was a typo from Jurgen and is in his "explained" file.
-The "log b" parameter in the "Midpoint" dataset actually appear to be "b" -- need to confirm this is the case.
-Species/variable/consensus group names are not consistent between all these data sources, but is handled in the data compiling below so there will be a single variable (species) that has the same grouping names between datasets

- Per Rob Pastorak- remove Corixidae species bc different feeding strategy from rest of group, so not using in our comparison- LM on 1/5/2022


## Individual Weights & Conversion Factors for Macroinvertebrates
```{r macro_LW}

## LW values for regression midpoints:
d.mid <- read.xlsx("Input/Macroinvertebrates_Biomass_Conversions.xlsx", sheet = "Macroinvert_Midpoints") %>% 
  janitor::clean_names() %>%
  mutate(species = ifelse(species == "Chironomus larvae", "Chironomidae", species)) # Change species name to match other data files -LM


## Mesocosm conversion factors - filter for macroinverts only
cf <- read.xlsx("//integral-corp.com/data/C2000-C3999/C3143_Mesocosm_Waterborne/Working_Files/R_Analyses/SpeciesList_ControlData_v_Aquatox.xlsx",
                  sheet = "Species List for Conversions") %>%
  janitor::clean_names() %>%
  filter(parameter_group == "Macroinvertebrates") %>%
  dplyr::select(parameter_group:resulting_units, units_for_aquatox, apply_species_lw_data) %>%
  dplyr::select(-criteria_midpoint) %>%
  filter(variable != "Corixidae") %>%
    # Add the weight per individuals to the conversion factor table- use the regression midpoint weights (group decision)
  mutate(weight_mg_per_ind = d.mid$weight_reg_mg_per_ind[match(.$apply_species_lw_data, d.mid$species)])

cf$variable[cf$variable=="Chironomidae (larvae)"] <- "Chironomidae" # rename to match control data merging later on

```

**Calculate Individual Weight (mg)- Lymnaeidae (<0.5cm)**
* There is no regression midpoint LW value for Lymnaeidae <0.5cm and the Lymnea (grown) range from 4-48cm, so not appropriate substitute
* Use the "Macroinvertebrate Populations" LW data from Jurgen/GmBH to find a weight per individual for Lymnaeidae

- Use parameters (a, b) from Jurgen's data
- Allometric equation: W = a x L^b
  - assuming W is body weight in mg
  - assuming L is body length in mm
  
- Coded equation: weight_mg_per_ind = (a*(median_length_mm^b))


```{r lymnaeidae_mass}
# 1. Read in Populations Data & LW values:
d.lym <- read.xlsx("Input/Macroinvertebrates_Biomass_Conversions.xlsx", sheet = "Macroinvert_All") %>%
  janitor::clean_names() %>%
  filter(species == "Lymnaeidae") %>%
  mutate(avg_length_mm = as.numeric(avg_length_mm)) %>%
  summarize(median_length_mm = median(avg_length_mm))

eq.lym <- read.xlsx("Input/Macroinvertebrates_Biomass_Conversions.xlsx", sheet = "Pop_Equation_Values") %>% 
  janitor::clean_names() %>%
  filter(species == "Lymnaeidae")

#2. Calculate the weight_mg_per_ind to add into the cf table:
lym <- bind_cols(eq.lym, d.lym) %>%
  mutate(weight_mg_per_ind = (a*(median_length_mm^b)))

#3. Add Lymnaeidae weight to cf table:
cf$weight_mg_per_ind[cf$variable=="Lymnaeidae (< 0.5 cm)"] <- lym$weight_mg_per_ind

rm(d.lym)
rm(eq.lym)
rm(lym)

```

## Control Mesocosm Data
* Start with the `sum_life_stages_bySpecies` dataset

```{r meso_control_data}

d.meso <- readRDS("Output/02_Control_and_Calib_byGroup/sum_life_stages_bySpecies.rds") %>% 
  filter(parameter_group=="Macroinvertebrates", variable != "Corixidae") %>%
  dplyr::select(study:date, abund_sum_lifestage)

```


# Convert Abundance to Biomass:
After joining the conversion factors to the control data...

1. Extrapolate sample abundance (ind/sample) to enclosure abundance (ind/L or ind/m2) ==> "enclosure_abundance"
    * Mesocosm control data reports counts as individuals per sample, using different sampling techniques depending on the critter.
    * The conversion factor, equations, and resulting units were provided by Waterborne.
    * Based on the conversion equations, Asellus and Gammarus will be ind/sample * (1+conversion factor), all others are ind/sample*conv factor

2. Calculate species-specific biomass
    * conversion methods will depend on the resulting_units (enclosure abundance) and the units_for_aquatox
    * Vol_to_area and area_to_vol ratios have been defined for cases when units change from volume to area and vice versa 
        * 1760 = volume of water (L) in the enclosures (i.e. total volume critters could live in)
        * 8.15 = area of enclosures (m2)
        
3. Sum biomass values by study, enclosure, date, and Aquatox record grouping (biomass_grouping)
    * We need to report 1 biomass value per study/enclosure/date/group.
    * Groupings are generally the same as consensus groups, with the exception for Odonates: we are modeling Anisoptera & Zygoptera separately


```{r convert_abund_to_biomass}
vol_to_area <- 1760/8.15
#area_to_vol <- 8.15/1760 #Note 2/4/22: When the Aquatox units were updated on 1/24, you no longer need the area_to_vol ratio.

d.bio <- left_join(d.meso, cf, by = c("parameter_group", "consensus_group", "variable")) %>%
  mutate(enclosure_abund = ifelse(consensus_group %in% c("Asellus", "Gammarus"), abund_sum_lifestage * (1 + conversion_factor),
                                    abund_sum_lifestage * conversion_factor),
         sp_biomass = case_when(
            resulting_units == "ind/L"  & units_for_aquatox == "mg/L dry" ~ enclosure_abund*weight_mg_per_ind,
            resulting_units == "ind/m2" & units_for_aquatox == "g/m2 dry" ~ (enclosure_abund*weight_mg_per_ind)/1000, # divide by 1000 to go from mg to g
            resulting_units == "ind/L"  & units_for_aquatox == "g/m2 dry" ~ (enclosure_abund*vol_to_area*weight_mg_per_ind)/1000), # divide by 1000 to go from mg to g
            #resulting_units == "ind/m2" & units_for_aquatox == "mg/L dry" ~ enclosure_abund*area_to_vol*weight_mg_per_ind),
         
         biomass_grouping = ifelse(apply_species_lw_data %in% c("Anisoptera", "Zygoptera"), apply_species_lw_data, consensus_group)) %>%
  group_by(study, enclosure, date, biomass_grouping) %>%
  mutate(group_biomass = sum(sp_biomass),
         group_abundance = sum(abund_sum_lifestage),
         group_encl_abundance = sum(enclosure_abund)) %>%
  distinct(study, enclosure, parameter_group, biomass_grouping, date, units_for_aquatox, group_biomass, group_abundance, group_encl_abundance)

write.csv(d.bio, paste0("Output/07_Biomass_Conversions/Macroinvertebrate_Biomass_by_ModeledGroups_", Sys.Date(), ".csv"))


```

```{r save_cf}
# Added 2/4/22-- Saving the compiled conversion factors & ratios to use in conversion of calibration criteria:

cf <- cf %>%
  mutate(vol_to_area = ifelse(resulting_units == "ind/L"  & units_for_aquatox == "g/m2 dry", vol_to_area, NA_real_),
         biomass_grouping = ifelse(apply_species_lw_data %in% c("Anisoptera", "Zygoptera"), apply_species_lw_data, consensus_group))

write_csv(cf, "Output/07_Biomass_Conversions/Macroinvertebrate_conv_factors.csv")
```


Subset table for Aquatox Overlay data file:
-12/14: B.Sackmann requested a table for Aquatox that has the species, original counts (ind/sample), extrapolated enclosure counts (ind/enclosure), and final biomass (g/m2 or mg/L)
```{r A_overlay}

# Output in format for AQUATOX observed values overlay:
AQ_MacroInvert <- d.bio %>%
  rename(species_group = biomass_grouping) %>%
  #dplyr::select(biomass_grouping, date, group_biomass, study, enclosure, abund_sum_lifestage, enclosure_abund, units_for_aquatox) %>%
  mutate(non_detects = 0, 
         min_bar_value = group_biomass,
         max_bar_value = group_biomass)
write.csv(AQ_MacroInvert, paste0("Output/07_Biomass_Conversions/Aquatox_Macroinvert_Overlay_", Sys.Date(), ".csv"))

```

#END QA





# Exploratory Methods - ARCHIVE - Not for QA
* We are no longer using both the population data, but keeping this information here in case we need to revert back to these methods - LM on 1/5/2021

1) Determine individual length for each species.
    - For species with population data, take the median of average values + IQR
    - For species without population data, use the midpoint values provided by Jurgen
    
2) Calculate individual mass (grams dry weight) using allometric equations from GmbH
    - Use the lengths determined above
    - Use the study specific a & b parameters (as there are different a & b values for species with both "midpoint" and "population" data)
    -Allometric equations return units of mg dry weight (based on figure legends), we convert to grams dry weight
    
3) Convert control data measurements from abundance (individuals/sample) to biomass (g/m2 or g/L)
    - Use the conversion factors and equations from Waterborne 

```{r old_data, eval=FALSE}
d.meso <- read_csv("//pfs1w/C2000-C3999/C3143_Mesocosm_Waterborne/Working_Files/R_Analyses/00_Model_Calibration/Output/02a_Control_data_by_Modeled_Groups/abund_by_modelgroups.csv") %>%
  filter(parameter_group == "Macroinvertebrates") %>%
  
      # Create a species column to match to the measured length data from Jurgen (using species in d.mid):
  mutate(species = case_when(
    variable == "Lymnaea stagnalis" ~ "Lymnaea",
    variable == "Notonecta glauca" ~ "Notonecta",
    variable == "Zygoptera" ~ "Zygoptera",
    variable == "Anisoptera" ~ "Anisoptera",
    variable == "Naididae" ~ "Naididae",
    #consensus_group %in% c("Asellus", "Chaoborus", "Chironomidae", "Cloeon", "Gammarus", "Tanypodinae", "Zygoptera", "Naididae") ~ consensus_group,
        consensus_group %in% unique(d.mid$species) ~ consensus_group,

    consensus_group == "Gastropoda" ~ variable
  )) #%>%
  #filter(!is.na(species)) # only keep records that have length data

#View(d.meso %>% distinct(consensus_group, variable, species))


# Get species list: 
  # pop.species = species with population level data, will use MEDIAN lengths for conversions
  # mid.species = species with only midpoint data, will use the MIDPOINT lengths for conversions
pop.species <- unique(d.pop$species)
mid.species <- d.mid %>% distinct(species) %>% filter(!species %in% pop.species) %>% pull(species) #pull turns it into a vector
```


```{r macro_pop_code, eval=FALSE}
## 1a: Read in the "populations" data, all column for date: NO LONGER USING 1/5/2022.
# d.pop <- read.xlsx("Input/Macroinvertebrates_Biomass_Conversions.xlsx", sheet = "Macroinvert_All") %>% 
#   janitor::clean_names() %>%
#   mutate(avg_length_mm = as.numeric(avg_length_mm),
#     date = case_when(
#     day_of_year == 183 ~ as.Date("2021-07-02"),
#     day_of_year == 232 ~ as.Date("2021-08-20"),
#     day_of_year == 261 ~ as.Date("2021-09-18")
#   ))
    # Source for determining date from day of year: https://www.esrl.noaa.gov/gmd/grad/neubrew/Calendar.jsp  

#d.sp <- unique(d$species)


## 1C: Read in the parameters for the allometric equations: NO LONGER USING 1/5/2022
# eq.pop <- read.xlsx("Input/Macroinvertebrates_Biomass_Conversions.xlsx", sheet = "Pop_Equation_Values") %>% 
#   janitor::clean_names() 

```


Steps 1 & 2: Determine length and calculate mass per individual:


For `pop.species` use the median lengths; for `mid.species` use the provided midpoints.

A. Conversions for groups with population level data  
```{r macro_pop_conversions}
d.pop.sum <- d.pop %>%
  group_by(species) %>%
  summarize(
    med_length_mm = median(avg_length_mm, na.rm=TRUE),
    q1_length_mm = quantile(avg_length_mm, 0.25, na.rm=TRUE),
    q3_length_mm = quantile(avg_length_mm, 0.75, na.rm=TRUE),
    IQR_length_mm = IQR(avg_length_mm, na.rm=TRUE)
  ) %>%
  ungroup()

## 4: Add the allometric parameters & convert length to biomass (into units of grams per individual)
    
d.pop.ind_mass <- left_join(d.pop.sum, eq.pop %>% dplyr::select(species, a, log10_a, b), by = "species") %>%
  # mutate(weight_g = exp((log10_a + b*med_length_mm))/1000, 
  #        q1_weight_g = exp((log10_a + b*q1_length_mm))/1000,
  #        q3_weight_g = exp((log10_a + b*q3_length_mm))/1000) 
  mutate(weight_g = (a*(med_length_mm^b))/1000,
         q1_weight_g = (a*(q1_length_mm^b))/1000,
         q3_weight_g = (a*(q3_length_mm^b))/1000)
```


Population length-weights, by date:
```{r macro_pop_conversions_by_date}

## 3: Get the median, 1st & 3rd quantile, and IQR for the average lengths-- by species & day measured:
d.popD.sum <- d.pop %>%
  group_by(species, day_of_year, date) %>%
  summarize(
    med_length_mm = median(avg_length_mm, na.rm=TRUE),
    q1_length_mm = quantile(avg_length_mm, 0.25, na.rm=TRUE),
    q3_length_mm = quantile(avg_length_mm, 0.75, na.rm=TRUE),
    IQR_length_mm = IQR(avg_length_mm, na.rm=TRUE)
  ) %>%
  ungroup()

## 4: Add the allometric parameters & convert length to biomass (into units of grams per individual)
    
d.popD.ind_mass <- left_join(d.popD.sum, eq.pop %>% dplyr::select(species, a, log10_a, b), by = "species") %>%
  # mutate(weight_g = exp((log10_a + b*med_length_mm))/1000, 
  #        q1_weight_g = exp((log10_a + b*q1_length_mm))/1000,
  #        q3_weight_g = exp((log10_a + b*q3_length_mm))/1000) 
  mutate(weight_g = (a*(med_length_mm^b))/1000,
         q1_weight_g = (a*(q1_length_mm^b))/1000,
         q3_weight_g = (a*(q3_length_mm^b))/1000) 

```

B. Conversions for groups with only midpoint data
12/13: Updated to use the length midpoint from the regression line, not the ranges given in the table. This is the method the other modeling groups will be using, per email conversations.
12/14: QA check- `the calc_weight_g` created here matches the values in email from Jurgen (orange table)
```{r macro_mid_conversions}

  
```

C. Compare individual masses for groups with population and midpoint data
-weight_g is individual mass from populations data
-calc_weight_g is individual mass from midpoints data
```{r}

d.mid.ind_mass <- left_join(d.mid, eq.mid %>% dplyr::select(species, a, log_a, b)) %>%
  mutate(calc_weight_g = (a*(length_reg_midpoint_mm^b))/1000)
combo.mass <- full_join(d.pop.ind_mass, d.mid.ind_mass, by = "species") %>%
  dplyr::select(species, calc_weight_g, weight_g, q1_weight_g, q3_weight_g)
  #mutate_if(is.numeric, signif, 3)
  

#Table for display in RMD:
cm_tab <- combo.mass %>%
  rename('Ind. Mass from Mid-point Lengths (g)'= calc_weight_g,
         'Ind. Mass from Pop. Lengths (g)' = weight_g,
         'Ind. Mass from Pop. Lengths-Q1 (g)' = q1_weight_g,
         'Ind. Mass from Pop. Lengths-Q3 (g)' = q3_weight_g)
#write.csv(cm_tab, "Output/07_MacroInvert_Conv/Calculated_Individual_Mass_g.csv")

pd <- combo.mass %>% 
  filter(!is.na(weight_g)) %>%
  pivot_longer(cols = c(calc_weight_g, weight_g), names_to = "method", values_to= "mass") %>%
  mutate(q1_weight_g = ifelse(method == "calc_weight_g", NA_real_, q1_weight_g),
         q3_weight_g = ifelse(method == "calc_weight_g", NA_real_, q3_weight_g))

# Plot of mass comparisons- only for those that have both
cm_plot <- ggplot(data=pd, aes(method)) +
  #geom_point(aes(y=weight_g)) +
  geom_pointrange(aes(y=mass, ymin= q1_weight_g, ymax=q3_weight_g)) +
  facet_wrap(~species, scales="free_y") +
  #theme(axis.text.x = element_blank()) +
  labs(x="", y= "Individual Biomass in Grams") +
  scale_x_discrete(labels= c("calc_weight_g"= "Regression Midpoint",
                             "weight_g" = "Population Data"))





# Plots of mass, calculated from populations data, by date:


popD_plot <- ggplot(d.popD.ind_mass, aes(x=date, y = weight_g)) +
  geom_point() +
  geom_line() +
  geom_ribbon(aes(ymin= q1_weight_g, ymax=q3_weight_g), alpha=0.5) +
  facet_wrap(~species, scales="free_y") +
  labs(x="Date of Measurement", y= "Individual Biomass (g) from Populations Data")
popD_plot

pdf(paste("Output/07_MacroInvert_Conv/Individual_Mass_By_Date_g.pdf",sep=""), paper="USr", width=10, height=7.25)
print(popD_plot)
dev.off()
```



Step 3: Convert control data measurements from abundance (individuals/sample) to biomass (g/m2 or g/L)

Notes:
-12/14: During internal call, decided that we will use the length values from the regressions (table d.mid and d.mid.ind_mass) as our starting length values for first determining individual biomass.
-12/14: B.Sackmann requested a table for Aquatox that has the species, original counts (ind/sample), extrapolated enclosure counts (ind/enclosure), and final biomass (g/m2)



```{r biomass_conversion, eval=FALSE}
#1. Subset the conversion factors table to only include columns needed:

cf_to_bind <- cf %>%
  dplyr::select(species, mesocosm_units, conversion_factor, conversion_equation, resulting_units)

#2. Subset the individual biomass data: Need in the individual biomass in grams dw per group
ind_mass_to_bind <- d.mid.ind_mass %>%
  dplyr::select(species, calc_weight_g) %>%
  rename(ind_mass_g = calc_weight_g) # Make the column name a bit more understandable - LM

# Below code used in exploratory analysis- this is a "mixed" approach, where if populations data was available then biomass from those data were used. Otherwise biomass calculated from the regressions were used. On 12/14 we decided not to go with this method since the other modeling teams were only using the regression biomass values - LM
# ind_mass_to_bind <- combo.mass %>%
#   mutate(ind_mass_g = ifelse(!is.na(weight_g), weight_g, calc_weight_g)) %>%
#   dplyr::select(species, ind_mass_g)
  

#3. Join the control data, individual biomass values, and conversion factors together. Calculate biomass & convert criteria from abundance to biomass.
d.meso.mass <- left_join(d.meso, ind_mass_to_bind, by = "species") %>%
  left_join(., cf_to_bind, by = "species") %>%
  mutate(enclosure_abund_ind_per_area = ifelse(species %in% c("Asellus", "Gammarus"), abund_sum_lifestage * (1 + conversion_factor),
                                    abund_sum_lifestage * conversion_factor),
         mass_g_per_area = ifelse(species %in% c("Asellus", "Gammarus"), abund_sum_lifestage * (1 + conversion_factor) * ind_mass_g,
                                    abund_sum_lifestage * conversion_factor * ind_mass_g),
         min_crit_mass_g = ifelse(species %in% c("Asellus", "Gammarus"), crit_min * (1 + conversion_factor) * ind_mass_g,
                                    min_peak_abund * conversion_factor * ind_mass_g),
         max_crit_mass_g = ifelse(species %in% c("Asellus", "Gammarus"), crit_max * (1 + conversion_factor) * ind_mass_g,
                                    min_peak_abund * conversion_factor * ind_mass_g),
         min_peak_mass_g = ifelse(species %in% c("Asellus", "Gammarus"), min_peak_abund * (1 + conversion_factor) * ind_mass_g,
                                    min_peak_abund * conversion_factor * ind_mass_g),
         max_peak_mass_g = ifelse(species %in% c("Asellus", "Gammarus"), max_peak_abund * (1 + conversion_factor) * ind_mass_g,
                                    max_peak_abund * conversion_factor * ind_mass_g),
         mass_units = gsub("ind", "g", resulting_units))



View(d.meso.mass %>% dplyr::select(species, abund_sum_lifestage, conversion_factor, enclosure_abund_ind_per_area, ind_mass_g, mass_g_per_area, min_peak_mass_g, max_peak_mass_g))

# Output of entire df:
write.csv(d.meso.mass, "Output/07_MacroInvert_Conv/MacroInvert_Control_Data_Converted_to_Biomass.csv")

# Output in format for AQUATOX observed values overlay:
AQ_MacroInvert <- d.meso.mass %>%
  dplyr::select(species, date, mass_g_per_area, study, enclosure, date2, abund_sum_lifestage, enclosure_abund_ind_per_area, mass_units) %>%
  mutate(final_units = paste0(mass_units, " dry")) %>%
  dplyr::select(-mass_units) %>%
  mutate(non_detects = 0, 
         min_bar_value = mass_g_per_area,
         max_bar_value = mass_g_per_area)
write.csv(AQ_MacroInvert, "Output/07_MacroInvert_Conv/Aquatox_Macroinvert_Overlay.csv")


## Create documentation output of species, conversion factors and equations used (requested by Waterborne)
macro.doc <- d.meso.mass %>%
  distinct(variable, .keep_all=TRUE) %>%
  dplyr::select(parameter_group, consensus_group, species, ind_mass_g, mesocosm_units, conversion_factor, conversion_equation, resulting_units, mass_units) %>%
  rename(conversion_sample_to_area = conversion_equation,
         resulting_units_sample_to_area = resulting_units,
         biomass_units = mass_units) %>%
  mutate(conversion_to_biomass = paste0("(", resulting_units_sample_to_area, ")", " * individual mass (g)")) %>%
  relocate(biomass_units, .after = conversion_to_biomass)

write_csv(macro.doc, "Output/07_MacroInvert_Conv/Macroinvertebrate_conversion_documentation.csv")
  
```
