==> AQUATOX/Code/R_Analyses/01_Model_Validation_Control/2021_Validation/11_Setup_Compare_Model_Validation_Criteria.Rmd <==
---
title: "Setup for Comparing Model & Validation Criteria"
author: "A. Lindborg"
date: "6/16/2022"
output: html_document
editor_options: 
  chunk_output_type: console
---
# Purpose
Complete any setup / test code to create a script for B.Sackmann that will take an Aquatox model output (.xls), grade model results against calibration criteria, and plot those together


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(readxl)
library(lubridate)
library(openxlsx)


# GGplot theme:
theme <- theme(text = element_text(family = "sans")) +
  theme_bw() +
  theme(legend.title = element_blank(), legend.text = element_text(color="black", size=8),
        legend.background = element_rect( size=.5, linetype= "solid", color = "black")) +
  theme(panel.grid.minor = element_blank(), panel.background = element_blank(), axis.line = element_line(colour = "black", size= 1)) +
  theme(axis.text.x = element_text(color="black", size=8, vjust=0.5, angle=0)) +
  theme(axis.text.y = element_text(color="black", size=8, vjust=0.5, angle=0)) +
  theme(axis.title.x = element_text(color="black", size=9, face="bold")) +
  theme(axis.title.y = element_text(color="black", size=9, face="bold")) +
  theme(plot.title = element_text(size = 10, hjust = 0.5, face="bold")) +
  theme(axis.ticks = element_line(size = 1)) +
  theme(axis.line = element_line(size = 1)) +
  theme(strip.background =element_rect(fill="white")) +
  theme(panel.border = element_rect(fill = NA, size = 0.25)) +
  theme(strip.text = element_text(face="bold", size = 9)) +
  theme(plot.margin = unit(c(0.5,0.5,.5,0.5), "cm"))

options(scipen = 9999)


# Create the output folder: SAVE THIS TIDBIT
# num_files <- length(list.files("Output"))
# run_no <- num_files + 1 
# dir.create(paste("Output/", run_no, "_Calibration_Results_", Sys.Date(), sep="")) # create the directory
# out.path <- paste("Output/", run_no, "_Calibration_Results_", Sys.Date(), sep="") # save the path for writing output

```

# Create inputs files:
- AQ model results
- Validation criteria in biomass
- Control data

* Save any modifications to inputs as new files in the validation folder for B.Sackmann
```{r inputs}

crit <- read_csv("Output/10_Convert_Calib_to_Biomass/Calibration_Criteria_in_Biomass.csv") %>%
  mutate(aq_record = ifelse(is.na(aq_record), consensus_group, aq_record)) %>% # No AQ records for total perio or total phyto- so fill in
  add_row(parameter_group = "Macroinvertebrates", consensus_group = "Odonata", aq_record = "Total Odonata", units_for_aquatox = "g/m2 dry",
          crit_biomass_min = .$crit_biomass_min[.$aq_record == "Dragonfly_Meso (g/m2 dry)"], 
          crit_biomass_max = .$crit_biomass_max[.$aq_record == "Dragonfly_Meso (g/m2 dry)"],
          .after = 10)

write_csv(crit, "Aquatox_Results_Validation_2021/Input/calibration_criteria.csv")

## Load Mesocosm data and add AQ record:
d.meso <- read_csv("Output/07_Biomass_Conversions/all_control_data_in_biomass.csv") %>% 
  mutate(year = lubridate::year(date)) %>%
  filter(parameter_group != "Macrophytes") %>% # don't need- no calibration criteria
  left_join(., crit%>%dplyr::select(parameter_group, consensus_group, aq_record)) %>%
  mutate(aq_record = case_when(
    species_group == "Anisoptera" ~ "Dragonfly_Meso (g/m2 dry)",
    species_group == "Zygoptera" ~ "Damselfly_Meso (g/m2 dry)",
    TRUE ~ aq_record
  )) %>%
  distinct() #somwhere along the way there were duplicates for the Odonates- this will remove them - LM 3/29/2022
  
  # Calculate biomass for total odonata and add to d.meso
d.od <- d.meso %>%
  filter(consensus_group == "Odonata") %>%
  group_by(study, enclosure, parameter_group, date, units_for_aquatox, consensus_group, year) %>%
  summarize(biomass = sum(biomass)) %>%
  mutate(aq_record = "Total Odonata", species_group = "Total Odonata")

d.meso <- bind_rows(d.meso, d.od)

write_csv(d.meso, "Aquatox_Results_Validation_2021/Input/mesocosm_control_data.csv")

# Need for filtering AQ data
exp.dates <- d.meso %>%
  #distinct(study, date) %>%
  group_by(study) %>%
  summarize(min_date = min(date),
            max_date = max(date))  %>%
  mutate(year = lubridate::year(min_date))
write_csv(exp.dates, "Aquatox_Results_Validation_2021/Input/experiment_dates.csv")


crit.study <- d.meso %>%
  group_by(parameter_group, consensus_group, study) %>%
 #Waterborne requested on 4/6/2022 that only the min and max should be used (not %5/95%)
  # summarize(min5 = quantile(biomass, 0.05),
 #           max95 = quantile(biomass, 0.95)) %>%
  summarize(min = min(biomass),
            max = max(biomass)) %>%
  mutate(year = case_when(
    study == "Full-Study-21a.1-2ApplS" ~ "2021_Application",
    study == "Full-Study-21a.2-3PeakS" ~ "2021_Peak"
  )) %>%
  dplyr::select(-study) %>%
  pivot_wider(names_from = year, values_from = c(min, max))


# add this to crit
crit.all <- left_join(crit, crit.study)
write_csv(crit.all, "Aquatox_Results_Validation_2021/Input/calibration_criteria_by_study_noquantile.csv")

crit.study.long <-d.meso %>%
  group_by(parameter_group, consensus_group, study) %>%
 #Waterborne requested on 4/6/2022 that only the min and max should be used (not %5/95%)
  # summarize(min5 = quantile(biomass, 0.05),
 #           max95 = quantile(biomass, 0.95)) %>%
  summarize(min = min(biomass),
            max = max(biomass)) %>%
  mutate(year = case_when(
    study == "Full-Study-21a.1-2ApplS" ~ "2021_Application",
    study == "Full-Study-21a.2-3PeakS" ~ "2021_Peak"
  )) %>%
  dplyr::select(-study)
write_csv(crit.study.long, "Aquatox_Results_Validation_2021/Input/calibration_criteria_by_study_long_noquantile.csv")

```




# ORIGINAL CODE- CAN BE OUT OF DATE, REFER TO 01_CALIBRATION_COMPARISONS.R FOR UPDATES AFTER 3/2/22

## Aquatox Results
```{r load_aq}
#### 3-2-2022: OUT OF DATE, REFER TO 01_ SCRIPT FOR REVISED CODE TO FIT NEW AQ RESULTS - LM
# Read in Aquatox model results:
files <- list.files(path = "AQUATOX_InitialCalibration_InProgress_IndYears_20220208",
                    pattern = "*.xls",
                    full.names = TRUE)

# Format AQ model results
aq.results <- tibble(filename = files) %>%
  mutate(contents = map(filename, ~read_excel(.x))) %>%
  unnest(cols=contents) %>%
  mutate(time = lubridate::hour(`...2`)) %>%
  filter(time ==12) %>% ## ONLY KEEP OBSERVATIONS FROM NOON
  dplyr::select(-c(`...2`, filename)) %>%
  rename(Date = `Date:`) %>%
  pivot_longer(cols= -c(Date),
               names_to = "aq_record",
               values_to = "model_result") %>%
  mutate(year = lubridate::year(Date)) %>%
  mutate(date.obj = as.Date(Date)) %>%
  filter(aq_record %in% unique(crit$aq_record)) %>%
  filter(date.obj >= exp.dates$min_date[match(year, exp.dates$year)] & date.obj <= exp.dates$max_date[match(year, exp.dates$year)]) %>%
  arrange(aq_record)

# Calculate total periphyton and total phytoplankton from AQ results:
aq.pp <- aq.results %>%
  filter(grepl("Peri", aq_record ) | grepl("Phyto", aq_record )) %>%
  mutate(aq_record = ifelse(grepl("Peri", aq_record), "Total Periphyton", "Total Phytoplankton")) %>%
  group_by(aq_record, date.obj, year, Date) %>%
  summarize(model_result = sum(model_result)) # checked both groups for first day in jan against totals from d.aq -all good - LM

aq.results <- bind_rows(aq.results, aq.pp)

# QA: aq.results %>% group_by(year) %>% summarize(min_date = min(date.obj), max_date = max(date.obj))
    # Verified that for each year, the mix and max dates match the experiment dates

rm(aq.pp)  # remove intermediary table
```


## Define study specific calibration criteria
* For each mesocosm control study, want to know the 5th percentile (min) and 95 percentile (max) -- point here is to see if the model results are modeling that experiment well, using the same setup for how overall calibration criteria created

*Integral methods: want to see how model compares to the actual control data*
```{r integral_methods}


# All years together:
int.grade <-  aq.results %>%
  left_join(., crit.all) %>%
  #left_join(., crit%>%dplyr::select(aq_record, crit_biomass_min, crit_biomass_max)) %>%
  mutate(meets_criteria = ifelse(model_result>= crit_biomass_min & model_result<=crit_biomass_max, TRUE, FALSE)) %>%
         #meets_study_crit = ifelse(model_result>= min5_2014 & model_result<=max95_2014, TRUE, FALSE)) %>%

  group_by(aq_record) %>%
  summarize(n_tot = n(),
            n_meet_crit = sum(meets_criteria),
            per_meet_crit = round((n_meet_crit/n_tot)*100,2)) %>%
            #n_meet_study = sum(meets_study_crit),
            #per_meet_study = (n_meet_study/n_tot)*100)%>%
  arrange(desc(n_meet_crit))


## Get the study-specific "criteria" met:
full.grade <- int.grade

for(i in c(2014, 2016, 2018, 2019)){
  
  year.crit <- crit.study.long %>% filter(year == i) %>% left_join(., crit)
  
  d <- aq.results %>% 
    filter(year == i) %>%
    left_join(., year.crit) %>%
    mutate(meets_criteria = ifelse(model_result>= crit_biomass_min & model_result<=crit_biomass_max, TRUE, FALSE),
           meets_study = ifelse(model_result>= min5 & model_result<= max95, TRUE, FALSE)) %>%
    group_by(aq_record) %>%
    summarize(n_tot = n(),
              n_meet_crit = sum(meets_criteria),
              per_meet_crit = round((n_meet_crit/n_tot)*100,2),
              n_meet_study = sum(meets_study),
              per_meet_study = round((n_meet_study/n_tot)*100, 2)) %>%
    arrange(desc(n_meet_crit)) %>%
    dplyr::select(-n_tot, -n_meet_crit, -n_meet_study)
  
  names(d) <- c("aq_record", paste0("per_meet_crit_", i), paste("per_meet_study_", i))
  
  full.grade <- left_join(full.grade, d)
}

names(full.grade) <- c("Aquatox Record", 
                       "Total Number of Modeled Points (all years combined)", 
                       "No. Modeled Points that Meet Criteria", 
                       "% Points Meet Calibration Criteria", 
                       "2014: % Criteria met", 
                       "2014: % Within Study Specific Ranges (5-95 percentiles)",
                       "2016: % Criteria met", 
                       "2016: % Within Study Specific Ranges (5-95 percentiles)",
                       "2018: % Criteria met", 
                       "2018: % Within Study Specific Ranges (5-95 percentiles)",
                       "2019: % Criteria met", 
                       "2019: % Within Study Specific Ranges (5-95 percentiles)")

# Save Workbook:
wb <- createWorkbook()
header_style <- createStyle(wrapText = TRUE)

addWorksheet(wb, sheetName= "Integral Calibration Check")
writeData(wb, sheet=1, x=full.grade , colNames = TRUE)
setColWidths(wb, sheet = 1, widths = 16, cols= 1:ncol(full.grade))
addStyle(wb, sheet=1, header_style, cols=1:ncol(full.grade), rows=1)

saveWorkbook(wb, paste0("Output/11_setup/test_IntegralTable.xlsx"), overwrite = TRUE)

#write.csv(full.grade, "Output/11_setup/Calibration_Scores_02092022.csv")
```

## Grading in Waterborne format - added 2/17
* Keeping code above as draft in case teams wants to keep seeing data the way we developed the tables
```{r wb_results_create_table1}

# Create grading table:
wb.grade <-  aq.results %>%
  left_join(., crit) %>%
  mutate(across(.cols = c(model_result, crit_biomass_min_peak, crit_biomass_max_peak, crit_biomass_min, crit_biomass_max), ~signif(.x, digits=2))) %>%
  #mutate(across(.cols = c(model_result, crit_biomass_min_peak, crit_biomass_max_peak, crit_biomass_min, crit_biomass_max), ~signif(.x, digits=3))) %>%
#year(wb.grade$peak_date_end) <- wb.grade$year
#year(wb.grade$peak_date_start) <- wb.grade$year

#wb.grade <- wb.grade %>%
  group_by(aq_record, year) %>%
  mutate(peak_value = ifelse(!is.na(peak_date_start), max(model_result), NA_real_),
         peak_date = if_else(!is.na(peak_date_start), min(date.obj[model_result == peak_value]), NA_real_), #Take first peak date that occurs
         max_value = ifelse(is.na(peak_date_start), max(model_result), # if there is no peak date, then just return the absolute max
                            max(model_result[date.obj<peak_date_start | date.obj>peak_date_end])), # if there is a peak date, return the max for the observations outside of the peak window
         #met_peak_date = ifelse(peak_date >= peak_date_start & peak_date <= peak_date_end, TRUE, FALSE)
         min_value = min(model_result),
         mean_value = mean(model_result)
         ) %>%
  ungroup() %>%
  distinct(aq_record, consensus_group, year, .keep_all = TRUE) %>%
    # Extact just the month/day for the peak windows & peak date
  mutate(start_monthday = as.numeric(format(peak_date_start, "%m%d")),
         end_monthday = as.numeric(format(peak_date_end, "%m%d")),
         peak_monthday = as.numeric(format(peak_date, "%m%d")))  %>%
  dplyr::select(-c(Date, model_result, date.obj, peak_date_start, peak_date_end, peak_date, units_for_aquatox)) %>%
  relocate(parameter_group, consensus_group, .before = aq_record) %>%
  relocate(start_monthday, end_monthday, crit_biomass_min_peak, crit_biomass_max_peak, crit_biomass_min, crit_biomass_max, peak_monthday, peak_value, min_value, max_value, .before=mean_value) %>%
  mutate(meet_peak = ifelse(is.na(peak_monthday), NA, 
                            ifelse(peak_monthday>=start_monthday & peak_monthday<=end_monthday & peak_value>=crit_biomass_min_peak & peak_value<= crit_biomass_max_peak, 1, 0)),
         meet_min = ifelse(min_value>=crit_biomass_min, 1, 0),
         meet_max = ifelse(max_value<=crit_biomass_max, 1, 0)) %>%
  group_by(aq_record) %>%
  mutate(peak_fulfilled = ifelse(sum(meet_peak) ==4, TRUE, FALSE),
         min_fulfilled = ifelse(sum(meet_min)==4, TRUE, FALSE),
         max_fulfilled = ifelse(sum(meet_max)== 4, TRUE, FALSE)) %>%
  ungroup() %>%
  pivot_wider(names_from = year, values_from = c("peak_monthday","peak_value" ,"min_value", "max_value","mean_value", "meet_peak", "meet_min", "meet_max")) %>%
  relocate(peak_value_2014, min_value_2014, max_value_2014, mean_value_2014,
           peak_monthday_2016, peak_value_2016, min_value_2016, max_value_2016, mean_value_2016,
           peak_monthday_2018, peak_value_2018, min_value_2018, max_value_2018, mean_value_2018,
           peak_monthday_2019, peak_value_2019, min_value_2019, max_value_2019, mean_value_2019, .after = peak_monthday_2014)


```

```{r wb_results_table1}
wb <- createWorkbook()
n <- 0
pos_style <- createStyle(fontColour = "#006100", bgFill = "#C6EFCE")

# Write out Biomass Data -- 28 species (as of 1/17/2022)
for(i in unique(wb.grade$parameter_group)) {
  
  d <- wb.grade %>%
    filter(parameter_group == i) %>%
    dplyr::select(consensus_group:crit_biomass_max, peak_monthday_2014:mean_value_2019) %>%
    dplyr::select(where(~!all(is.na(.)))) %>%
    arrange(consensus_group)
  
  n_rows <- nrow(d)
  n_cols <- ncol(d) # peri & phyto will be 17, macro & zoo will be 29 bc have peak info
  
  addWorksheet(wb, sheetName= i)
  writeData(wb, sheet=i, x=d , colNames = TRUE)
  
  ## Add conditional formatting to color cellls

  
}

saveWorkbook(wb, paste0("Output/11_setup/test_table1.xlsx"), overwrite = TRUE)



## highlight cells in column 1 based on value in column 2
# writeData(wb, "Dependent on", data.frame(x = 1:10, y = runif(10)), startRow = 15)
# conditionalFormatting(wb, "Dependent on",
#   cols = 1,
#   rows = 16:25, rule = "B16<0.5", style = negStyle
# )
# conditionalFormatting(wb, "Dependent on",
#   cols = 1,
#   rows = 16:25, rule = "B16>=0.5", style = posStyle
# )
```

```{r wb_results_table2}

# Creat the "Y/N" Criteria fulfilled table
wb.grade.yn <- wb.grade %>%
  dplyr::select(parameter_group:crit_biomass_max, peak_fulfilled, min_fulfilled, max_fulfilled)


wb.grade.perc <- wb.grade.yn %>%
  pivot_longer(cols = c(peak_fulfilled, min_fulfilled, max_fulfilled), names_to = "criteria_type", values_to = "criteria_result") %>%
  filter(!is.na(criteria_result)) %>%
  summarize(total_crit = n(),
            total_crit_met = sum(criteria_result),
            percent_met = total_crit_met/total_crit * 100) %>%
  dplyr::select(percent_met) %>%
  rename(`Percent Criteria Met (all studies/groups)` = percent_met)

wb <- createWorkbook()

neg_style <- createStyle(bgFill = "#FF0000") # red background
pos_style <- createStyle(bgFill = "#CCFFCC") # green background
#centerStyle <- createStyle(halign = "center")


for(i in unique(wb.grade.yn$parameter_group)) {
  
  d <- wb.grade.yn %>%
    filter(parameter_group == i) %>%
    mutate(peak_fulfilled = ifelse(peak_fulfilled==TRUE, "Y", "N"),
           min_fulfilled = ifelse(min_fulfilled==TRUE, "Y", "N"),
           max_fulfilled = ifelse(max_fulfilled ==TRUE, "Y", "N")) %>% 
    rename(`Species Group` = consensus_group,
           `AQUATOX record` = aq_record,
           `Earliest peak date` = start_monthday,
           `Latest peak date` = end_monthday,
           `Minimum Peak Biomass` = crit_biomass_min_peak,
           `Maximum Peak Biomass` = crit_biomass_max_peak,
           `Minimum Biomass` = crit_biomass_min,
           `Maximum Biomass` = crit_biomass_max,
           `Peak Date and Biomass` = peak_fulfilled,
           `Minimum Biomass Fulfilled` = min_fulfilled,
           `Maximum Biomass Fulfilled` = max_fulfilled) %>%
    dplyr::select(where(~!all(is.na(.)))) %>%
    dplyr::select(-parameter_group)
  
  n_rows <- nrow(d) +1 #add 1 so when writes out, will include the last row
  n_cols <- ncol(d) # peri & phyto will be 17, macro & zoo will be 29 bc have peak info
  
  addWorksheet(wb, sheetName= i)
  
  writeData(wb, sheet=i, x=d , colNames = TRUE)
  setColWidths(wb, sheet = i, cols = 1:n_cols, widths = "auto")
  
  conditionalFormatting(wb, sheet = i,
  cols = 1:n_cols,
  rows = 1:n_rows, 
  rule = '="Y"', 
  style = pos_style)
  
  conditionalFormatting(wb, sheet = i,
  cols = 1:n_cols,
  rows = 1:n_rows, rule = '="N"', 
  style = neg_style)
  
}

writeData(wb, sheet = "Macroinvertebrates", x = wb.grade.perc, startCol= 13)

conditionalFormatting(wb, sheet = "Macroinvertebrates",
  cols = 13,
  rows = 2,
  rule = ">=90",
  style = pos_style)

conditionalFormatting(wb, sheet = "Macroinvertebrates",
  cols = 13,
  rows = 2,
  rule = "<90",
  style = neg_style)


saveWorkbook(wb, paste0("Output/11_setup/test_tableYN.xlsx"), overwrite = TRUE)


# percentage met, not included peak criteria- just for exploratory purposes
# wb.grade.perc.np <- wb.grade.yn %>%
#   pivot_longer(cols = c(peak_fulfilled, min_fulfilled, max_fulfilled), names_to = "criteria_type", values_to = "criteria_result") %>%
#   filter(!is.na(criteria_result) & criteria_type != "peak_fulfilled") %>%
#   summarize(total_crit = n(),
#             total_crit_met = sum(criteria_result),
#             percent_met = total_crit_met/total_crit * 100)


```

```{r wb_plots}

### Adjust timescale for plotting
aq.results$date2 <- aq.results$date.obj
year(aq.results$date2) <- 2021

d.meso$date2 <- d.meso$date
year(d.meso$date2) <- 2021

exp.dates.p <- exp.dates
year(exp.dates.p$min_date) <- 2021
year(exp.dates.p$max_date) <-2021

# Set min and max dates for X-axis, as requested by B.Sackmann:
min_date <- as.Date("2021-05-10")
max_date <- as.Date("2021-09-30")

### Print plots
pdf(paste0(out.path, "/", run_no,  "_Calibration_Result_Plots.pdf", sep=""), paper="USr", width=10, height=7.25)
par(omi=c(1.5,0.5,0,0.5), pin=c(7,8),pty='m')

for(i in unique(d.meso$parameter_group)){
  pd <- d.meso %>% filter(parameter_group == i) %>% arrange(aq_record)
  
  for(j in unique(pd$aq_record)){

  md <- pd %>% filter(aq_record == j)
  cd <- aq.results %>% filter(aq_record == j)
  units <- regmatches(j, gregexpr("(?=\\().*?(?<=\\))", j, perl=T))[[1]]
  
  p <- ggplot() +
    geom_point(data = md, aes(x=date2, y = biomass), color = "darkgreen") +
    geom_path(data=cd, aes(x=date2,  y=model_result), size=1, color= "black") +
    geom_vline(data = exp.dates.p, aes(xintercept = min_date), color= "blue") +
    geom_vline(data = exp.dates.p, aes(xintercept = max_date), color= "blue") +
    facet_wrap(~year) +
    geom_hline(yintercept = crit$crit_biomass_min[crit$aq_record == j], color= "darkgray", linetype="dotted") +
    geom_hline(yintercept = crit$crit_biomass_max[crit$aq_record == j], color= "darkgray", linetype="dotted") +
    geom_vline(xintercept= crit$peak_date_start[crit$aq_record == j], color="orange") +
    geom_vline(xintercept= crit$peak_date_end[crit$aq_record == j], color="orange") +
    labs(x= "", y= paste0("Biomass ", units), title = paste0("Parameter Group: ", i, "\nConsensus Group: ", unique(md$consensus_group), "\nAQUATOX record: ", j)) +
    scale_x_date(limits=c(min_date, max_date)) +
    theme
  print(p)
  }
}
dev.off()



```




# DRAFT CODE- ARCHIVE
```{r draft_data_wrangling}

# Need for plotting:
crit.dates <- crit %>%
  filter(!is.na(peak_date_start)) %>%
  dplyr::select(parameter_group, consensus_group, aq_record, peak_date_start, peak_date_end)

d.aq <- read_excel("//pfs1w/C2000-C3999/C3143_Mesocosm_Waterborne/Working_Files/R_Analyses/00_Model_Calibration/AQUATOX_InitialCalibration_InProgress_IndYears_20220208/005_ASM_RingStudy_Mesocosm_Model - Ext - ICPrepComb - 7yrSUM_Export_C_2014.xls") %>%
  #dplyr::select(-2) %>%
  mutate(time = lubridate::hour(`...2`)) %>%
  filter(time ==12) %>% ## ONLY KEEP OBSERVATIONS FROM NOON
  dplyr::select(-`...2`) %>%
  rename(Date = `Date:`) %>%
  pivot_longer(cols= -c(Date),
               names_to = "aq_record",
               values_to = "model_result") %>%
  mutate(year = lubridate::year(Date)) %>%
  mutate(date = as.Date(Date))



# Calculate total periphyton and total phytoplankton from AQ results:
d.pp <- d.aq %>%
  filter(grepl("Peri", aq_record ) | grepl("Phyto", aq_record )) %>%
  mutate(aq_record = ifelse(grepl("Peri", aq_record), "Total Periphyton", "Total Phytoplankton")) %>%
  group_by(aq_record, date, year, Date) %>%
  summarize(model_result = sum(model_result)) # checked both groups for first day in jan against totals from d.aq -all good - LM

d.aq <- bind_rows(d.aq, d.pp)

```

```{r draft_plot_code}
# draft a plot with Asellus
da <- d.aq.clean %>% filter(aq_record == "Asellus_Meso (g/m2 dry)")

ggplot(da, aes(x=date, y=model_result)) +
  geom_path() +
  facet_wrap(~aq_record) +
  geom_point(data = d.meso %>% filter(year==2014 & consensus_group == "Asellus"), aes(x=date, y = biomass)) +
  geom_hline(yintercept = crit.all$crit_biomass_min[crit.all$aq_record == "Asellus_Meso (g/m2 dry)"], color= "orange") +
  geom_hline(yintercept = crit.all$crit_biomass_max[crit.all$aq_record == "Asellus_Meso (g/m2 dry)"], color= "orange")



# Same as above, with Chaoborus
dc <- d.aq.clean %>% filter(aq_record == "Chaoborus_Meso (mg/L dry)")

ggplot(dc, aes(x=date, y=model_result)) +
  geom_path() +
  facet_wrap(~aq_record) +
  geom_point(data = d.meso %>% filter(year==2014 & consensus_group == "Chaoborus"), aes(x=date, y = biomass)) +
  geom_hline(yintercept = crit.all$crit_biomass_min[crit.all$aq_record == "Chaoborus_Meso (mg/L dry)"], color= "orange") +
  geom_hline(yintercept = crit.all$crit_biomass_max[crit.all$aq_record == "Chaoborus_Meso (mg/L dry)"], color= "orange")


# Example of Chaoborus- not meeting any criteria
test.c <- test %>% filter(aq_record == "Chaoborus_Meso (mg/L dry)")

ggplot() +
  geom_path(data=test.c %>% filter(year ==2014), aes(x=date.obj, y=model_result)) +
  geom_path(data=test.c %>% filter(year ==2016), aes(x=date.obj, y=model_result)) +
  geom_path(data=test.c %>% filter(year ==2018), aes(x=date.obj, y=model_result)) +
  geom_path(data=test.c %>% filter(year ==2019), aes(x=date.obj, y=model_result)) +
  #facet_wrap(~aq_record) +
  geom_point(data = d.meso %>% filter(consensus_group == "Chaoborus"), aes(x=date, y = biomass)) +
  geom_hline(yintercept = crit.all$crit_biomass_min[crit.all$aq_record == "Chaoborus_Meso (mg/L dry)"], color= "orange") +
  geom_hline(yintercept = crit.all$crit_biomass_max[crit.all$aq_record == "Chaoborus_Meso (mg/L dry)"], color= "orange")


## Plots for all critters - compared against the "main" calibration criteria only. Show all 4 model years:
pdf(paste("Output/11_setup/calibration_plots.pdf",sep=""), paper="USr", width=10, height=7.25)
par(omi=c(1.5,0.5,0,0.5), pin=c(7,8),pty='m')
for(i in unique(test$aq_record)){
  
  test.c <- test %>% filter(aq_record == i) #%>% separate(col=aq_record, into = c("species", "units"), sep = "\\(", extra="merge", remove=FALSE)
  #grades <- full.grade %>% filter(aq_record == i) ## Use this to add calibration grade info
  units <- regmatches(i, gregexpr("(?=\\().*?(?<=\\))", i, perl=T))[[1]]
  
  p <- ggplot() +
    geom_point(data = d.meso %>% filter(aq_record == i), aes(x=date, y = biomass), color = "grey45") +
    geom_path(data=test.c %>% filter(year ==2014), aes(x=date.obj, y=model_result), size=1, color= "deeppink1") +
    geom_path(data=test.c %>% filter(year ==2016), aes(x=date.obj, y=model_result), size=1, color= "deeppink1") +
    geom_path(data=test.c %>% filter(year ==2018), aes(x=date.obj, y=model_result), size=1, color= "deeppink1") +
    geom_path(data=test.c %>% filter(year ==2019), aes(x=date.obj, y=model_result), size=1, color= "deeppink1") +
    #facet_wrap(~aq_record) +
    geom_hline(yintercept = crit.all$crit_biomass_min[crit.all$aq_record == i], color= "orange") +
    geom_hline(yintercept = crit.all$crit_biomass_max[crit.all$aq_record == i], color= "orange") +
    labs(x= "", y= paste0("Biomass ", units), title = i) +
    theme
  
  print(p)
}
dev.off()


```
